Load editor icons in editor structs
Took 10 minutes
