+++ crates/dream-editor/src/assets_panel.rs
@@ -1,5 +1,8 @@
 use egui_wgpu::Renderer;
 
+use dream_renderer::image::Image;
+use dream_renderer::texture;
+
 use crate::editor::Panel;
 
 pub struct AssetsPanel {
@@ -12,15 +15,45 @@ impl AssetsPanel {
         renderer: &dream_renderer::renderer::RendererWgpu,
         egui_wgpu_renderer: &mut Renderer,
     ) -> Self {
+        let file_icon_texture_bytes = include_bytes!("icons/FileIcon.png");
+        let mut file_icon_image = Image::default();
+        file_icon_image.load_from_bytes(file_icon_texture_bytes, "icons/FileIcon.png", None);
+        let rgba = file_icon_image.to_rgba8();
+        let file_icon_texture = texture::Texture::new(
+            &renderer.device,
+            &renderer.queue,
+            rgba.to_vec(),
+            rgba.dimensions(),
+            None,
+        )
+        .expect("Unable to load file icon texture");
+
+        let directory_icon_texture_bytes = include_bytes!("icons/DirectoryIcon.png");
+        let mut directory_icon_image = Image::default();
+        directory_icon_image.load_from_bytes(
+            directory_icon_texture_bytes,
+            "icons/DirectoryIcon.png",
+            None,
+        );
+        let rgba = directory_icon_image.to_rgba8();
+        let directory_icon_texture = texture::Texture::new(
+            &renderer.device,
+            &renderer.queue,
+            rgba.to_vec(),
+            rgba.dimensions(),
+            None,
+        )
+        .expect("Unable to load directory icon texture");
+
         let file_epaint_texture_id = egui_wgpu_renderer.register_native_texture(
             &renderer.device,
-            &renderer.file_icon_texture.view,
+            &file_icon_texture.view,
             wgpu::FilterMode::Linear,
         );
 
         let directory_epaint_texture_id = egui_wgpu_renderer.register_native_texture(
             &renderer.device,
-            &renderer.directory_icon_texture.view,
+            &directory_icon_texture.view,
             wgpu::FilterMode::Linear,
         );
