+++ crates/dream-editor/src/icons/CollapseIcon.png
