+++ crates/dream-editor/src/icons/DirectoryIcon.png
