+++ crates/dream-editor/src/icons/EditIconDark.png
