+++ crates/dream-editor/src/icons/EditIconLight.png
