+++ crates/dream-editor/src/icons/ExpandIcon.png
