+++ crates/dream-editor/src/icons/FileIcon.png
