+++ crates/dream-editor/src/icons/PlayIcon.png
