+++ crates/dream-editor/src/icons/PlayIconDark.png
