+++ crates/dream-editor/src/icons/PlayIconLight.png
