+++ crates/dream-editor/src/icons/SelectIconBlack.png
