+++ crates/dream-editor/src/icons/SelectIconWhite.png
