+++ crates/dream-editor/src/icons/StopIcon.png
