+++ crates/dream-editor/src/icons/WrenchIconDark.png
