+++ crates/dream-editor/src/icons/WrenchIconLight.png
