+++ crates/dream-editor/src/renderer_controls_panel.rs
@@ -1,6 +1,9 @@
 use egui::Widget;
 use egui_wgpu::Renderer;
 
+use dream_renderer::image::Image;
+use dream_renderer::texture;
+
 use crate::editor::Panel;
 
 pub struct RendererControlsPanel {
@@ -12,9 +15,22 @@ impl RendererControlsPanel {
         renderer: &dream_renderer::renderer::RendererWgpu,
         egui_wgpu_renderer: &mut Renderer,
     ) -> Self {
+        let play_icon_texture_bytes = include_bytes!("icons/PlayIcon.png");
+        let mut play_icon_image = Image::default();
+        play_icon_image.load_from_bytes(play_icon_texture_bytes, "icons/PlayIcon.png", None);
+        let rgba = play_icon_image.to_rgba8();
+        let play_icon_texture = texture::Texture::new(
+            &renderer.device,
+            &renderer.queue,
+            rgba.to_vec(),
+            rgba.dimensions(),
+            None,
+        )
+        .expect("Unable to load play icon texture");
+
         let play_icon_epaint_texture_id = egui_wgpu_renderer.register_native_texture(
             &renderer.device,
-            &renderer.play_icon_texture.view,
+            &play_icon_texture.view,
             wgpu::FilterMode::Linear,
         );
